# static-ui/demo/demo/openapi.yaml
openapi: 3.0.3
info:
  title: Static UI Bug Tracker API
  version: 0.1.0
  description: >
    Simplified API contract for backend implementation.
    Covers auth, users, workspaces, bugs, comments, attachments, dashboard.
servers:
  - url: http://localhost:8080
...
security:
  - bearerAuth: []
paths:
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login and get JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginReq'
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRes'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register user and return JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserReq'
      responses:
        '200':
          description: Register success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRes'
        '400':
          description: Validation or duplicate email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/bugs/{id}:
    get:
      tags: [Bugs]
      summary: Get bug detail
      parameters:
        - $ref: '#/components/parameters/BugId'
      responses:
        '200':
          description: Bug detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BugDetailRes'
        '404':
          description: Bug not found
    put:
      tags: [Bugs]
      summary: Update bug
      parameters:
        - $ref: '#/components/parameters/BugId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBugReq'
      responses:
        '200':
          description: Bug updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bug'
    delete:
      tags: [Bugs]
      summary: Delete bug
      parameters:
        - $ref: '#/components/parameters/BugId'
      responses:
        '204':
          description: Bug deleted
        '409':
          description: Delete rule violation
...
  /api/bugs/{bugId}/comments:
    get:
      tags: [Comments]
      summary: List comments by bug
      parameters:
        - $ref: '#/components/parameters/BugIdPathNamed'
      responses:
        '200':
          description: Comment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
    post:
      tags: [Comments]
      summary: Add comment to bug
      parameters:
        - $ref: '#/components/parameters/BugIdPathNamed'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
...
  /api/bugs/{bugId}/attachments:
    get:
      tags: [Attachments]
      summary: List attachments by bug
      parameters:
        - $ref: '#/components/parameters/BugIdPathNamed'
      responses:
        '200':
          description: Attachment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Attachment'
    post:
      tags: [Attachments]
      summary: Upload attachment for bug
      parameters:
        - $ref: '#/components/parameters/BugIdPathNamed'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Attachment uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attachment'
  /api/attachments/{id}/download:
    get:
      tags: [Attachments]
      summary: Download attachment by id
      parameters:
        - $ref: '#/components/parameters/AttachmentId'
      responses:
        '200':
          description: File stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Attachment not found
  /api/attachments/{id}:
    delete:
      tags: [Attachments]
      summary: Delete attachment by id
      parameters:
        - $ref: '#/components/parameters/AttachmentId'
      responses:
        '204':
          description: Attachment deleted
  /api/dashboard/summary:
    get:
      tags: [Dashboard]
      summary: Summary count by status
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
      responses:
        '200':
          description: Status summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  open:
                    type: integer
                  inprogress:
                    type: integer
                  resolved:
                    type: integer
                  closed:
                    type: integer
  /api/dashboard/by-priority:
    get:
      tags: [Dashboard]
      summary: Bug count by priority
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
      responses:
        '200':
          description: Priority summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  high:
                    type: integer
                  medium:
                    type: integer
                  low:
                    type: integer
  /api/dashboard/by-developer:
    get:
      tags: [Dashboard]
      summary: Bug count by assignee
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdQuery'
      responses:
        '200':
          description: Assignee summary
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    assigneeId:
                      type: integer
                      nullable: true
                    assigneeName:
                      type: string
                    count:
                      type: integer
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    WorkspaceIdQuery:
      name: workspaceId
      in: query
      required: false
      schema:
        type: integer
    BugId:
      name: id
      in: path
      required: true
      schema:
        type: integer
    BugIdPathNamed:
      name: bugId
      in: path
      required: true
      schema:
        type: integer
...
    AttachmentId:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: BAD_REQUEST
        message:
          type: string
          example: Validation failed
    LoginReq:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 4
    CreateUserReq:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          example: Nguyen Van A
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 4
        role:
          type: string
          example: developer
        phone:
          type: string
          example: "0912345678"
...
    UserDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
        phone:
          type: string
          nullable: true
        avatar:
          type: string
          nullable: true
    AuthRes:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/UserDto'
...
    Bug:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
          example: BUG-10
        title:
          type: string
        description:
          type: string
          nullable: true
        priority:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [open, inprogress, resolved, closed]
        reporterId:
          type: integer
        assigneeId:
          type: integer
          nullable: true
        workspaceId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
...
    UpdateBugReq:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [open, inprogress, resolved, closed]
        assigneeId:
          type: integer
          nullable: true
    Comment:
      type: object
      properties:
        id:
          type: integer
        bugId:
          type: integer
        userId:
          type: integer
        text:
          type: string
        createdAt:
          type: string
          format: date-time
    Attachment:
      type: object
      properties:
        id:
          type: integer
        bugId:
          type: integer
        fileName:
          type: string
        fileSize:
          type: integer
          format: int64
        uploadedAt:
          type: string
          format: date-time
    BugDetailRes:
      allOf:
        - $ref: '#/components/schemas/Bug'
        - type: object
          properties:
            comments:
              type: array
              items:
                $ref: '#/components/schemas/Comment'
            attachments:
              type: array
              items:
                $ref: '#/components/schemas/Attachment'
...