# static-ui/demo/demo/src/main/java/com/example/demo/model/dto/AuthRes.java
package com.example.demo.model.dto;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class AuthRes {
    private String token;
    private UserDto user;
}