# static-ui/demo/demo/src/main/java/com/example/demo/model/dto/CommentDto.java
package com.example.demo.model.dto;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import java.time.LocalDateTime;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class CommentDto {
    private Integer id;
    private Integer bugId;
    private Integer userId;
    private String userName;
    private String userAvatar;
    private String text;
    private LocalDateTime createdAt;
}