# static-ui/demo/demo/src/main/java/com/example/demo/model/dto/UserDto.java
package com.example.demo.model.dto;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor

public class UserDto {
    private Integer id;

    private String name;

    private String email;

    private String role;

    private String phone;

    private String avatar;

}